+++ hiredis/__init__.py
@@ -1,6 +1,10 @@
-from .hiredis import Reader, HiredisError, ProtocolError, ReplyError
+from .hiredis import Reader, HiredisError, pack_command, ProtocolError, ReplyError
 from .version import __version__
 
 __all__ = [
-  "Reader", "HiredisError", "ProtocolError", "ReplyError",
+  "Reader",
+  "HiredisError",
+  "pack_command",
+  "ProtocolError",
+  "ReplyError",
   "__version__"]